Fix time calculation to be cumulative time instead of per-stage time
LiveSplit expects the game time returned from here to be the total time
so far for this run, not just the current stage's time. Therefore, keep
track of the previous stages's time in a variable, and add to that.

In addition, assume that restarting the current level doesn't reset the
run (so add the time spent to the total time, as per Speedrun.com's
rules), and that restarting the tutorial level or switching to the
tutorial level does reset the run (so reset the total time to 0).

It mostly works, but I'm sure there's a few cases where it breaks/get
messy.

Signed-off-by: Saikrishna Arcot <saiarcot895@gmail.com>